requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
sentence-transformers>=2.2.0
accelerate>=0.25.0

# Agent framework
langgraph>=0.0.40
langchain>=0.1.0
langchain-community>=0.0.10

# Vector store
faiss-cpu>=1.7.4

# NLP utilities
spacy>=3.7.0
# After install run: python -m spacy download en_core_web_sm

# LLM inference
bitsandbytes>=0.41.0      # 4-bit quantization for Mistral-7B

# Data & config
pyyaml>=6.0
pandas>=2.1.0

# Evaluation
rouge-score>=0.1.2
scikit-learn>=1.3.0

# Document processing
pymupdf>=1.23.0           # PDF text extraction
python-docx>=1.0.0        # DOCX read/write (Stage 1 input + Stage 4 report output)
docx2pdf>=0.1.8           # Stage 4 report DOCX → PDF (uses LibreOffice headless on Linux)

# Experiment tracking
wandb>=0.16.0

# Testing
pytest>=7.4.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
